% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hotpot}
\alias{hotpot}
\title{Hotpot learner}
\usage{
hotpot(
  X,
  y,
  scale = T,
  boundary_function = boundary_half_circle,
  lower = c(theta = -pi),
  upper = c(theta = pi)
)
}
\description{
Hotpot learner
}
